vpp: fix dpdk plugin dlopen symbol errors by switching to system dpdk
vpp build-system supports static linking of dpdk, or dynamic linking to
the system-wide installed dpdk libraries.

During static linking dependencies such as libpcap and zlib are not
picked up by the build system.

The use-system-dpdk flag enabels using pkg-config and dynamic linking,
which together work perfectly as long as the individual dpdk libs are
present in the rootfs.

With yocto the standard method works well and avoids aforementioned
symbol errors.